eve-alpine: remove stage 0 step for go 1.20.1
In order to update to go 1.20.1 lines to the dockerfile were added
to download go from edge/community from alpine.
Unfortunately the package for aarch64 has been removed, but fortunately
it is still in our /mirror directory. So the solution is to
base this image on the eve-alpine image that has an up-to-date
/mirror directory and remove downloading go version 1.20.1 and
rely only on /mirror

Signed-off-by: Christoph Ostarek <christoph@zededa.com>