pkg: newlog: Do not crash on a bad timestamp
Formatted log messages (coming from memlogd, for instance) are suppose to
have a timestamp in the RFC 3339 standard. In this format, dates can be
represented through different ways. For example, the strings below
represent the same date:

2020-12-31T21:07:14-05:00
2021-01-01T02:07:14Z

The Z character in the second one means UTC time. In getMemlogMsg(), if
ParseLogInfo() was not able to get the timestamp, the one used in the
original message will be used. However, it assumes only the second format
(with suffix Z). This commit removes this assumption and let the timestamp
to be parsed in getPtypeTimeStamp(). If it fails, it considers Unix default
(initial) time as the timestamp.

Signed-off-by: Renê de Souza Pinto <rene@renesp.com.br>