pkg: newlog: Do not parse application's console as JSON log messages
The output console from an application VM can be logged, i.e., have
contents sent from memlogd to newlogd container. In this case, it should
not be parsed as a JSON log message.

This commit removes the calling of ParseLogInfo() for any log message
coming from an application and treat it as a normal log message. Also, it
searches for "guest_vm" only in the field source of the log message.
Otherwise, any JSON log message containing the string "guest_vm" would be
considered as an application's log message.

Signed-off-by: Renê de Souza Pinto <rene@renesp.com.br>