zedrouter: Fix metadata server for switch network instances
createServer4 expects bridge IP address without prefix length,
but this wasn't the case with switch network instances.
This (rather old) bug was introduced because we used 'string' instead
of 'net.IP' for the data type and therefore it wasn't clear what is
the expected value. In order to avoid the confusion, I have also changed
the type to net.IP. An additional bonus is that we avoid some
net.ParseIP calls repeatedly made across zedrouter.

Signed-off-by: Milan Lenco <milan@zededa.com>