Revert "kernels,storage-init,mkimage-raw-efi: make ZFS built-in"
This partially reverts commit 17dd29bfdd5356204afdccec6fdc4be81a447df2.

With this revert we build ZFS again as a module. Why?
The linux kernel is licensed GPL, but OpenZFS is licensed CDDL.
These two are incompatible. So we compile both (kernel and ZFS)
and ship them separately.

This patch breaks collecting of the kernel dumps for the /persist
volume formatted as ZFS. We still jump into the kdump and output
the dmesg to the console, but not dump or logs are stored in the
/persist volume. This will be fixed in future releases of the
EVE-OS

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>