debug/collect-info.sh: don't copy folders, create symlinks instead
This presumably fixes the #3118 by creating symlinks of folder
with huge files (e.g. logs) instead of copying them to the tmpfs
folder, which can cause the OOM.

This patch also installs GNU tar for ignoring all warnings and
read errors.

Bump up the version of the script to v2.

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>