client: Try to fetch the UUID each boot.
previously, we would declare success if the old uuid existed, but there
is a possibility that a device has been deleted and re-onboarded, so
attempt to actually fetch the UUID and re-register.

This allows you to recover an accidently deleted device by rebooting
instead of requiring a re-image.

Helps with #3109

Signed-off-by: Harsh Modi <hmodi@aviatrix.com>