NI state: Handle inactive VIF + minor improvements
When application VIF is inactivated, Network Instance State Collector
should stop collecting state data for the interface and also it should
forget previously recorded IP assignements. Once the interface is
reactivated, state collecting is resumed and the newly leased IP
address(es) will be recorded and an update will be sent to zedrouter
(even if re-activated VIF obtains the same IP as before).
This scenario happens for example during an application purge.

This commit additional includes renaming of some variables related
to metrics for more clarity.

Lastly, the method collecting ACL counters was incorrectly including the
OUTPUT chain, which, however, is not being used and as a result an error
message was frequently logged.

Signed-off-by: Milan Lenco <milan@zededa.com>