zedcloud/deferred: don't take lock while processing the deferred queue
The ctx.lock protects the concurrent access to the ctx.deferredItems,
so no need to keep the mutex locked while processing the queue.

Why? Queue processing can take up to several minutes (depends on the
send timeout), so that all other goroutines which produce new requests
(callers of the SetDeferred() or RemoveDeferred() calls) will stuck
for as long as the queue is not completely processed. This mutex
contention badly impacts overall system responsiveness.

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>