zedmanager: don't require restart field set in case of FixedResources.DisableLogs field change
The default behavior of the quantifyChanges() is to return 'restart == true'
in case of any change in the `FixedResources` structure (type VmConfig).

The field `DisableLogs` is handled without Application restart requirement,
thus pull the field out of the `VmConfig` structure (which is used as
'fixed resources') and move it in the parent structure.

This fixes the bug, when 'Disable logs' field is changed on the controller
side but EVE returns an error: "Need restart due to FixedResources changed".

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>