Move HTTP send operations away from the zedagent main event loop
Zedagent main event loop is responsible for handling pubsub updates along
with sending messages to the controller. HTTP send is synchronous call
and can stall the whole main event loop for up to 6 minutes (overall send
timeout).
This badly affects the system responsiveness which can be observed in
air-gaped environment (no connectivity to the controller), when the queue
is stuck waiting for send to be completed and no other pubsub updates
are handled.

Two things are done in this commit:

* Introduced connect() ("dial" in terms of Go) timeout, which is much less
  than the send() timeout (helps to fail faster).
* all direct HTTP send() calls are moved away from the main event loop
  and executed in dedicated goroutines.

Signed-off-by: Milan Lenco <milan@zededa.com>