arm: Add support to install EVE image on eMMC for i.MX8MP devices
This commit enables the generation of installer raw image for i.MX8MP
devices. This image can be flashed onto an SD Card that will be able to
boot and execute the normal installation process and install EVE on the
eMMC of the running device.

A small patch for U-Boot is provided in order to set mmcdev u-boot variable
automatically according to the boot device (SD Card or eMMC) for EPC-R3720
platform, as it's already done in Pollux platform. In this way there is no
need to change this variable across SD Card or eMMC images.

Signed-off-by: Renê de Souza Pinto <rene@renesp.com.br>