hypervisor/kvm: reduce VMM overhead to 100MiB for containers
The whole motivation of this patch is to run more containers-in-vm
applications on the EVE node, reducing the VMM max memory ("overhead").
Having that container's load is not so intensive the VMM max memory
is reduced to 100MiB.

If customer needs higher VMM limit - `VMMMaxMem` configuration can
be set per App basis.

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>