pillar/volumemgr: Implement snapshot-related functions in VolumeHandler.
This commit adds the necessary functions to the VolumeHandler struct for
handling snapshots. The functions include CreateSnapshot,
RollbackToSnapshot, and DeleteSnapshot. These functions enable the
creation, rollback, and deletion of snapshots for file-based volumes.
They also mock the implementation for container and YFS based volumes.

The data returned by low-level snapshot handlers is stored as an
interface metadata. This approach allows for flexibility as the specific
data returned depends on the underlying file system.

For file-based volumes, the CreateSnapshot function returns the snapshot
name used by the qemu-img tool as metadata.

Signed-off-by: Nikolay Martyanov <nikolay@zededa.com>