pillar/volumemgr: Add pub/sub interface for snapshot configuration and status messages.
This commit introduces a pub/sub interface for handling snapshot-related
messages.

The commit introduces a new subscription, subVolumesSnapConfig,
responsible for receiving VolumesSnapshotConfig messages. It includes
create, modify, and delete handlers to process changes related to
snapshot configuration.

Additionally, a new publication, pubVolumesSnapStatus, is created to
publish VolumesSnapshotStatus messages. This publication enables the
dissemination of snapshot status updates, which will be sent as a
response to zedmanager.

In the types/volumetypes.go file, the commit introduces the
VolumesSnapshotConfig struct. This struct represents a snapshot request
from zedmanager to volumemgr and includes fields such as SnapshotID,
Action, VolumeIDs, AppUUID, and ConfigID. It also provides a Key()
function to generate a unique key for the snapshot.

Furthermore, the commit adds a new struct called VolumesSnapshotStatus.
This struct captures the status information of snapshots and includes
relevant attributes to track the progress and state of snapshots. The
VolumesSnapshotStatus messages will serve as responses from volumemgr to
zedmanager, providing updates on the snapshot progress.

Signed-off-by: Nikolay Martyanov <nikolay@zededa.com>