Make domainmgr skip DNS with empty DPCKey
Do not activate PhysicalIOAdapterList subscription until NIM receives
a DPC and publishes corresponding DNS.
NIM can publish DNS even before it receives first DPC. In such case
DPCKey is empty.
The goal is to avoid assigning network ports to PCIBack if they are going to be
used for management purposes. This way we avoid doing unintended port assignment
to PCIBack that would be shortly followed by a release, therefore mitigating
the risk of race conditions between domainmgr and NIM.

Signed-off-by: Milan Lenco <milan@zededa.com>
Signed-off-by: yash-zededa <yash@zededa.com>