spec.sh: enumerate cellular modems with USB addr included
Now that EVE supports multiple cellular modems, it is preferred to use
USB addresses to reference modems from the device model. This is because
interface names are not deterministic and depend on the order of modem
initialization, which may differ on each boot.
It is therefore desirable for spec.sh to enumerate cellular modems with
the USB address included.

Moreover, assigngrp is now properly set to avoid USB controller used for
a modem to be accidentally assigned to an application.

Lastly, to make it easier for users to identify multiple attached modems,
spec.sh obtains description for each modem using Hayes command ATI.

Signed-off-by: Milan Lenco <milan@zededa.com>