Separate radio silence config from status
Config stores the intended state of the radio silence mode,
while status stores the actual state.

Previously, if the operation to enable radio silence failed,
RadioSilence.Imposed would be set to false and because the same field
was used for both status and the intended config, RS flag passed
to wwan microservice would be inadvertently reverted back to false.
As a consequence, after the initial unsuccessful attempt to enable RS,
the wwan microservice would cease making any further retry attempts.

Signed-off-by: Milan Lenco <milan@zededa.com>