pillar/zedmanager: Fix rollback error reporting.
When EVE fails to initiate a rollback, ensure the error is correctly set
for proper propagation to the status sent to the controller. Additional
details like time and severity have been included.

Also, set the time of the error during setting snapshot error in the
setSnapshotStatusError function if it's not set by the source of the
error.

Signed-off-by: Nikolay Martyanov <nikolay@zededa.com>