pillar/zedmanager: Pass SnapshotID directly to deserializeConfigFromSnapshot.
Updated the function deserializeConfigFromSnapshot to directly take the
SnapshotID string as an argument instead of the whole
SnapshotInstanceStatus object. This change enables the function to be
used in scenarios where the full Status object is not available, but the
SnapshotID is known. Function calls and handling of snapshot directory
were updated accordingly to align with this change.

Signed-off-by: Nikolay Martyanov <nikolay@zededa.com>