pillar/zedmanager: Snapshot metadata serialization for rollback post-restart.
Introduced functions to serialize and deserialize snapshot metadata,
crucial for enabling rollback functionality after a system restart.

The serializeSnapshotMetadata function serializes SnapshotInstanceStatus
into a JSON file that is saved in the snapshot directory under the
SnapshotMetadataFilename. This enables persistent storage of snapshot
metadata, ensuring it is not lost during a system restart.

The deserializeSnapshotMetadata function reads the snapshot metadata
JSON file and converts it back into a SnapshotInstanceStatus instance.
This allows the application to retrieve stored snapshot metadata
following a restart.

The restoreAvailableSnapshots function is implemented to read through
snapshot directories and restore the metadata for each snapshot
belonging to a specific application. This function is invoked in
handleCreate to ensure that any available snapshots are restored during
the application creation process after a system restart.

A SnapshotMetadataFilename constant has been introduced in
locationconsts.go to specify the filename used for storing snapshot
metadata.

Signed-off-by: Nikolay Martyanov <nikolay@zededa.com>