pillar/zedmanager: Enhance modularity of snapshot data structure restoration.
This commit restructures the handling of snapshots for improved
modularity and extensibility. The key changes include:

1. Introduction of a new function, `addFixupsIntoSnappedConfig`, in
   `handlevolumemgr.go`. This function adds fixups to the snapped app
   instance config. These are elements that should be drawn from the
   current app instance config, rather than the snapshot.

2. Refactoring of the `triggerRollback` function in `updatestatus.go`.
   Restoration of `volumeRefStatuses` from the snapshot config is now
   handled by the new function `restorePresnapStatus`.

These updates have been made with a view to facilitate future additions
of restore functions. By separating the restoration of different data
structures into dedicated functions, the code is more modular, easier to
read, and can be extended more efficiently.

Signed-off-by: Nikolay Martyanov <nikolay@zededa.com>