zedrouter/appcontainer: fix connectivity to the VM docker
From some version of the docker/client library the .NewClient()
call was deprecated, which caused the following error:

  "getAppContainers: Container list error Cannot connect to the Docker daemon at tcp://$IP:2375. Is the docker daemon running?"

because by default the connectivity was established through
the unix docker.sock socket.

This patch fixes the problem by passing the HTTP client to
the constructor of the docker client.

The main motiviation of this fix is the "Unknown" state of
"Modules" visible on the UI of the controller, because
statistics was never collected from the docker.

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>