pillar/volumemgr: Prevent fatal error for preexisting Volume Status.
The Volume Status is locally reestablished during rollback. Following this, if
the controller dispatches a new config featuring the rolled-back volume, it
prompts an unclearly triggered handleDeferredVolumeCreate function. Notably,
the preexistence of volume status will no longer yield a fatal error; it
instead triggers a warning and an early function return.

Signed-off-by: Nikolay Martyanov <nikolay@zededa.com>