NI Reconciler: Add item representing IP address reservation
IPReserve is an item representing allocation and use of an IP address
(for bridge). The purpose of this item is to ensure that the same IP
address will not be used by multiple bridges at the same time
(this includes intermediate reconciliation states).
This works by having the bridge depending on the reservation and
by requesting re-creation of IPReserve when it changes, thus triggering
re-create of bridges and all higher-layers items that depend on it.

Signed-off-by: Milan Lenco <milan@zededa.com>