Handle new memlogd format
The format changes in the new version we pulled in PR #3293 resulted in
producing log entries in /persist/newlog/collect/ which are effectively
empty. The change in memlogd as seen by logread is a common vs a a
semicolon but the underlying change is to produce json which we need to
parse in newlogd.

Signed-off-by: eriknordmark <erik@zededa.com>