pillar/zedmanager: Add local AppInstanceConfig subscription.
Enhanced `lookupAppInstanceConfig` function to include an additional boolean
parameter that checks local AppInstanceConfig when set to true. Also, added
new function `lookupLocalAppInstanceConfig` to retrieve local
AppInstanceConfig.

Added subscription and publication for local AppInstanceConfig in the
zedmanager context. This serves as a placeholder for a config copy that Eve
reverts to when necessary. So, this subscriptions is checked first now. It
includes logic to use local AppInstanceConfig in `checkDelayedStartApps` and
`handleModify`.

Updated all calls to `lookupAppInstanceConfig` in multiple files to include the
new boolean parameter.

Implemented `unpublishLocalAppInstanceConfig` and
`publishLocalAppInstanceConfig` for managing local AppInstanceConfig
publication.

Signed-off-by: Nikolay Martyanov <nikolay@zededa.com>