pillar/zedmanager: Rework the rollback flow.
In updatestatus.go, the check for RollbackInProgress has been removed from the
updateAIStatusUUID and doUpdate functions, thus not skipping the update process
during a rollback. We should not skip now, as the config version is taken from
a special local subscription that provides us with the proper version. The
check for RollbackInProgress has also been removed from the doInstall function.
Instead, a new check for RollbackInProgress is added after checking if an app
purge is in progress. This means the system will not activate an application if
a rollback is still in progress.

In zedmanager.go, the triggering of the rollback has been moved to the
handleModify function from the doUpdate function. This means a rollback request
can now be handled directly when the app configuration is modified. The
initiation of the rollback now sets HasRollbackRequest and RollbackInProgress
to true. The rollback also publishes the application instance status and
updates the configuration to the one from the snapshot.

The handleModify function also has changes to the purge handling. The needPurge
condition now has one more option: if HasRollbackRequest is true. This is to
ensure that a purge occurs during a rollback (we need it to upgrade app
instance volumes).

The rollback handling that was previously in handleModify has been removed, as
it's now handled earlier in the same function.

In zedmanagertypes.go, the comments for HasRollbackRequest and
RollbackInProgress in the SnapshottingStatus struct have been updated to
reflect the new handling of rollbacks. The ConfigBeforeRollback field has been
removed because the rollback flow is now controlled by the RollbackInProgress
field, and the version of the configuration available in the module is the
correct one, not the previous one. Hence, it doesn't need to store the previous
configuration version.

Signed-off-by: Nikolay Martyanov <nikolay@zededa.com>