pillar/zedmanager: Rework rollback trigger to apply old configuration.
This commit significantly revises the triggerRollback function within the
Zedmanager's updatestatus module. Instead of restoring VolumeRefStatuses and
the AppInstanceConfig from a snapshot, the function now creates or updates a
VolumesSnapshotConfig to indicate a rollback action. If a VolumesSnapshotConfig
is found for the snapshot in question, its action is set to rollback. If no
matching VolumesSnapshotConfig is found, a new one is created with the relevant
VolumeIDs from the current AppInstanceStatus, and the action set to rollback.
This change simplifies the rollback process and makes it more resilient to
system reboots.

Signed-off-by: Nikolay Martyanov <nikolay@zededa.com>