pillar/volumemgr: Rework snapshot configuration handlers.
This commit introduces significant changes to the handling of volume snapshots
in our system. Key changes include:

In volumemgr.go, the handlers for creating, modifying, and deleting volume
snapshots have been reworked. The new handlers are
`handleVolumesSnapshotConfigCreate`, `handleVolumesSnapshotConfigModify`, and
`handleVolumesSnapshotConfigDelete`. These handlers now call a common
implementation function `handleVolumesSnapshotConfigImpl`.

In handlesnapshot.go, the functions `handleVolumesSnapshotCreate`,
`handleVolumesSnapshotModify`, and `handleVolumesSnapshotDelete` have been
significantly reworked. They now call `handleVolumesSnapshotConfigImpl` with
the appropriate action set in the snapshot config. The function
`handleVolumesSnapshotConfigImpl` handles the creation, rollback, and deletion
of snapshots based on the action in the config.

New functions `createSnapshot`, `rollbackSnapshot`, and `deleteSnapshot` have
been introduced in handlesnapshot.go to handle the respective actions.

The functions `publishVolumesSnapshotStatus` and
`unpublishVolumesSnapshotStatus` remain largely unchanged, but the function
`lookupVolumesSnapshotStatus` has been updated to check if the snapshot status
exists in the store if it is not found in the published state. It's necessary
to be able to do rollback after the system reboot. For that a new function
`deserializeVolumesSnapshotStatus` has been introduced to read the snapshot
status from a file.

In updatestatus.go, the function `triggerSnapshotDeletion` has been updated to
set the `Action` of `volumesSnapshotConfig` to `types.VolumesSnapshotDelete`
before unpublishing the snapshot config. It's an optional change.

Signed-off-by: Nikolay Martyanov <nikolay@zededa.com>