pillar/zedagent: Refactor config processing flags to distinguish reboot and update scenarios.
In order to better handle different scenarios where configuration processing
needs to be skipped, the `configProcessingSkipFlag` in the `getconfigContext`
struct was refactored into a new type `configProcessingRetval`. This new type
allows us to distinguish between cases where we want to skip configuration
processing due to a reboot (`skipConfigReboot`) or due to an update
(`skipConfigUpdate`). An additional `isSkip` method was also added to the
`configProcessingRetval` type for convenience.

The configuration processing code has been updated to use these new flags and
to correctly handle different scenarios. For example, we now handle cases where
we are in the middle of a baseOS upgrade and need to skip processing the app
instances.

Signed-off-by: Nikolay Martyanov <nikolay@zededa.com>