pillar/zedagent: Add string representation for configProcessingRetval and enhance logging.
A new method (String) has been added to the configProcessingRetval enumeration,
which provides string representations for its various states. This method aids
in better logging and debugging of the system.

Additionally, the log message in the configTimerTask function has been modified
to print the string representation of the configProcessingSkipFlag when it is
not OK, thereby providing more descriptive

Signed-off-by: Nikolay Martyanov <nikolay@zededa.com>