Makefile: Workaround to fix QEMU's MSR errors
A QEMU crash has being experienced for some combinations of newer kernels
and QEMU versions on the host when running EVE from QEMU (with make
run-live). QEMU crashes when trying to access VMX_PROCBASED_CTLS2 MSR
register and VMs/Containers are not initialized:

error: failed to set MSR 0x48b to 0x137bff00000000

This commit provides a workaround while the proper fix is not provided.

Signed-off-by: Renê de Souza Pinto <rene@renesp.com.br>