Generate random MAC addresses even for local network instances
Even for local network instances we should generate random MAC address
that with a high probability will not collide with other apps on other
edge nodes. From networking point of view this is not necessary,
but there are apps that use MAC address as some sort of ID and could
fail to function properly if there is a collision across edge nodes.

As opposed to switch network instance, here the collision could be
a problem for such apps even if edge nodes are in completely different
locations and connected to different networks. We therefore need even
higher probability of address uniqueness. For example, with 24 random
bits and 1000 app interfaces the change of a collision would be as high
as 3%. Since these MAC addresses will not appear on external Ethernet
networks, we can also use OUI octets for randomness. Only I/G and
U/L bits need to stay constant and set appropriately.

Signed-off-by: Milan Lenco <milan@zededa.com>