debug : fix collecting tpm event log
TPM event log is a kernel memory mapped to a pseudo-filesystem,
as a result when sized always returns 0. This is problematic for
tar because it won't read the file (symlink or not) and we end up
with an empty tpm event log in the final tar file.

This change fixes the issue by creating a copy of it as a real file,
the file size is limited 1mb.

Signed-off-by: Shahriyar Jalayeri <shahriyar@zededa.com>