attest : fix race-condition in tests
the attestation uses a state-machine to process the request,
the state-machine safely runs on a single go routine, but to test it we
need to externally set the state machine status to diffrent values to
simulate a request, this requires to spawn a seperate go routine, this
leads to read/writes to some internal values that causes data race-conditions.

this changes introduces atomic read/write helpers to fix the issue.

Signed-off-by: Shahriyar Jalayeri <shahriyar@zededa.com>