Avoid using cached controller IP for connectivity verification
This commit completely refactors the mechanism of caching the controller
IP address. Previously, the routine (run by NIM) that periodically looks
up controller IP would store it to /etc/hosts to basically cache it
and apply it for all processes running on EVE.
However, this for example then means that even connectivity verification
(as performed by DPCManager and UplinkProber) uses cached IP address and
doesn't always check the availability of the DNS service. This can be
an issue in various scenarious. For example, user may change device
config and set a wrong DNS IP address (inside static network IP config).
EVE will try this new config and because controller IP is cached from
a previous lookup, it will not check if DNS is working and may therefore
evaluate the new config as OK and forget the previous one. Once the TTL
of the cached IP runs out, EVE will finds itself in a situation where
DNS with the new IP is not responsive, controller hostname cannot be
resolved into IP and at the same time there is no network config to
fallback to. Thus device looses connectivity permanently.

Having cache applied to all processes also hinders troubleshooting
activities. Collected netdumps are likely not going to contain traces of
DNS requests because cached IP was used instead. Similarly, users that
manually will try to perform lookup to check how their DNS server behaves
may get tricked thinking that everything is working, when in fact IP
address stored in /etc/hosts was only returned. Output they collect
using tools like nslookup or dig is not going to be very useful if there
was no actual DNS query performed.

The purpose of the cache is to limit the number of DNS queries executed
over time. This is just an optimization and it should not hinder
essential functionality, such as network config fallback.
This commit changes controller IP caching to use pubsub for storage
and apply the cached IP only within pillar and only in situations where
we do no care if DNS server was actually queried. This is for example
in cmd/client, loguploader and zedagent. For NIM, diag and UplinkProber
we always want to run full DNS queries. They do not run them that often
anyway.

Signed-off-by: Milan Lenco <milan@zededa.com>
(cherry picked from commit 0af4043f715baa76fe2374786d01269b9d010006)