Support routing where gateway depends on link route
Especially in public clouds it is common to assign /32 IPs to VM
instances and publish a link route for the gateway via DHCP.
This is used to reduce ARP traffic, enhance security, isolation,
scalability of IP address allocations, etc.

In physical networks, this is rather unlikely configuration.
Still, since we test EVE on GCP, we must support routing where
gateway depends on a link route. For the routes added to per-NI
routing tables it means that we need to better describe route
dependency, because the gateway could be either inside the subnet
of the IP assigned to the output interface or it could be routed
by a link-scoped route (and this link route must be configured first).
Both cases must be supported.

Signed-off-by: Milan Lenco <milan@zededa.com>