pillar/zedagent: Implement NodeAgent readiness check in zedagent initialization.
This commit introduces a NodeAgent readiness check to ensure its status is
available before the zedagent initialization process. This is necessary to
guarantee the successful initialization of certain parts of the zedagent
context, specifically 'bootReason'. Without this check, there's a risk of these
components remaining uninitialized. The readiness check is done through a wait
loop which only progresses once the NodeAgent status is confirmed to be ready.

Signed-off-by: Nikolay Martyanov <nikolay@zededa.com>