CODEOWNERS: fix formatting
Apparently the CODEOWNERS format does not accept the multi-lines,
so if there are multiple reviewers for the same patters, those
should be specified on the same line. Here is what doc says:

"If you want to match two or more code owners with the same pattern,
 all the code owners must be on the same line. If the code owners are
 not on the same line, the pattern matches only the last mentioned
 code owner."

But who reads docs ...

There is no any difference in reviewers list, only format change.

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>