Update edgeview to be compatible with the refactored zedrouter
Some pillar/types on which Edgeview depends were changed during
the Zedrouter refactoring. Please note that these types
are not persisted, so we do not concern ourselves with backward
compatibility (as it would significantly limit development).
The Edgeview changes done in this commit only affects the server
side, meaning that users will not have to update the client-side
container.

Signed-off-by: Milan Lenco <milan@zededa.com>