pillar: Refactor deserialization logic into utils package.
This commit refactors the deserialization logic used in volumemgr and
zedmanager packages into a new utility function, utils.DeserializeToStruct, to
reduce code duplication and improve readability. The redundant code in the
handleSnapshot and handlevolumemgr files is replaced with this new utility
function. In addition, a new utility file deserialize.go is introduced to
contain the shared deserialization logic. The related imports are also updated
accordingly.

Signed-off-by: Nikolay Martyanov <nikolay@zededa.com>